[update] add the timeout about change result
